pyproject.toml
[project]
name = "agent-genome-watcher"
version = "0.1.0"
description = "Observing How AI Societies Consume Questions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "playwright>=1.58.0",
    "openai>=2.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
